chore: clean up event unmarshalling, remove async API calls (#38)
The cloudevents sdk has some inconsistencies with how it exposed the
unmarshalling. Version 0.3 and earlier had an async API, while 1.0
(smartly) changed that. As a result, the impl in this file was pretty
ugly. By using the receivers directly, we can avoid that and clean up
this code.

These changes are based on: https://github.com/cloudevents/sdk-javascript/pull/120

Once those changes land in a released version of the cloudevents sdk
this file can be even simpler as the sdk will handle most of this work.

As an aside - dropped support for 0.2. Normally I wouldn't be so casual about
dropping support, but upstream sdk has no 0.2 support in the next release
and cloud events spec says only the two most recent versions are required
to be supported.