deps: bump to fastify 3.3.0 (#51)
Changes in 3.2.1 to fastify resulted in a broken framework. The fastify
module added a `context` property to the request object, which we had
been decorating with a property by the same name. This commit bumps to
the latest 3.3.0 of fastify, and changes the property that we decorate
the request object with to `fcontext`. These changes are entirely
internal and do not affect the user function.