fix: reverse the parameter order when invoking
To make it possible for functions to easily act as both HTTP and CloudEvent
handler it's better to have any potential data as the second parameter.

This change also makes use of a new bit of the cloudevents API to receive an
incoming event, eliminating a transitive dependency on axios.

Fixes: https://github.com/boson-project/faas-js-runtime/issues/62

Fixes: https://github.com/boson-project/faas/issues/194 (when updated in thetemplate)

Signed-off-by: Lance Ball <lball@redhat.com>