feat: add TypeScript type definitions
This commit adds type definitions for the following:
  * `Invokable`: The user provided function
  * `Context`: The invocation context
  * `CloudEventResponse`: Convenience class for constructing events
  * `Logger`: The logging interface
  * `LogLevel`: Logging enumeration
  * `start`: A function to start the invoker
  * `InvokerOptions`: Options for `start`

The definitions are statically tested with `tsd`.

These type definitions are not actually used by the JS files. This does not
convert the module to TypeScript; it just adds these type definitions to our
exports. This will allow us to create TypeScript function templates without
a full conversion of this code base to TypeScript. Ultimately, I think it's
probably best to make the change.

This change also bumps to `cloudevents@4.0.2` which does have a breaking
change. See: https://github.com/cloudevents/sdk-javascript/issues/332
As a result, we accept all versions of CloudEvents. Actually, I think this
is probably a fine change in the long run. Loose validation upon receipt
of an event allows functions to handle event sources that may be producing
malformed events. I've commented out the test for now. If we decide it makes
sense to leave this as-is instead of waiting for a fix from upstream, then
we should remove the test and the existing logic for dealing with this,
since it's no longer necessary.

Finally, I've really simplified the `start` funtion interface. It now just
takes a function and a bag of options. It doesn't deal with callbacks anymore.
Promises are used to deal with async server startup.

Signed-off-by: Lance Ball <lball@redhat.com>