feat: add limited support for func.yaml via logLevel
This commit introduces an `options.config` parameter to `start()` which
may point to a directory or a `func.yaml` file. When the server starts,
this config file will be read, if it exists, to determine the expected
log level.

Note that this determines the __defaults__ upon server start up. These
may be overridden via the options parameter as usual. E.g. calling
`start(func, { logLevel: 'trace' })` will set the log level to `trace`
no matter what's in `func.yaml`.

Signed-off-by: Lance Ball <lball@redhat.com>