chore: move publish to its own workflow and add sigstore provenance
This pulls the actual publication of the module into its own workflow,
which is run when a release is created. It also adds sigstore provenance
to the publication so that the release is signed by sigstore and logged
in a public transparency ledger.

See: https://docs.npmjs.com/generating-provenance-statements
See: https://www.sigstore.dev/

Signed-off-by: Lance Ball <lball@redhat.com>