chore: change to using rome linter instead of eslint
The `eslint` linter has become large and unwieldly with frequent updates
and too many transitive dependencies. This PR changes the linter to
Rome (https://rome.tools/). Rome can handle both formatting and linting,
is very fast, and requires minimal configuration.

I've added a single configuration option in `rome.json` which specifies
using single quotes, which I'm not married to, but it seems pointless to
change the quote style everywhere. I'm happy to add additional linting
or formatting rules to reduce the number of changes with this PR, but
honestly, I'm happy to go with what's already the default if that
doesn't cause this PR to be too disruptive.

Signed-off-by: Lance Ball <lball@redhat.com>